+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
-## v0.1.0-dev
+## v0.1.0
 
 * Initial version
+    * Struct definition
+    * Type definition
+    * Default values
+    * Enforced keys